Move env var to init
